Add files via upload
Here in Mountain View, you get a pretty diverse Asian community, such as Asian linkage restaurants, supermarkets where you can get ethnic Asian food, Asian street food, and beverage shops, etc. Given the fast-paced situation of jobs changing, I might be taking an offer on the other side of the state but which has a similar community nearby. 

This is how I came up with the idea of clustering other cities in the state which shares identical community and cultures. The information revealed by this report would be useful for new immigrants to California who is looking to the same diversity environments.

Because it helps you to locate the spot where you should be living to gain similar neighborhoods where you have lived.